requirements.txt
pip install Django==2.0.3
pip install djangorestframework
djangorestframework-jwt
